refactor: interfaceをrepositoryに定義
