+++ internal/infra/mysql/cluster.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 
+	"github.com/ryo-funaba/example_echo/internal/domain/repository"
 	"github.com/ryo-funaba/example_echo/internal/utils/enum"
 	"github.com/ryo-funaba/example_echo/internal/utils/errorutil"
 	"github.com/volatiletech/sqlboiler/v4/boil"
@@ -18,15 +19,7 @@ type cluster struct {
 	isDebugMode bool
 }
 
-type Cluster interface {
-	GetConnection(conn enum.ConnectionType, database string) (*sql.DB, error)
-	AddConnection(conn enum.ConnectionType, database, driver, dns string) error
-	AddConnectionForTest(conn enum.ConnectionType, database, driver, dns string) error
-	GetBoilCtxExecutor(ctx context.Context, conn enum.ConnectionType, database string) (boil.ContextExecutor, error)
-	SetTxInCtx(ctx context.Context, tx *sql.Tx) context.Context
-}
-
-func NewCluster(isDebugMode bool) Cluster {
+func NewCluster(isDebugMode bool) repository.Cluster {
 	pools := map[string]*sql.DB{}
 	c := &cluster{pools, isDebugMode}
